Modstrings: Working with modified nucleotide sequences

Representing nucleotide modifications in a nucleotide sequence is usually done via special characters from a number of sources. This represents a challenge to work with in R and the Biostrings package. The Modstrings package implements this functionallity for RNA and DNA sequences containing modified nucleotides by translating the character internally in order to work with the infrastructure of the Biostrings package. For this the ModRNAString and ModDNAString classes and derivates and functions to construct and modify these objects despite the encoding issues are implemenented. In addition the conversion from sequences to list like location information (and the reverse operation) is implemented as well.

See also

Author

Maintainer: Felix G.M. Ernst felix.gm.ernst@outlook.com (ORCID)

Other contributors:

Denis L.J. Lafontaine denis.lafontaine@ulb.ac.be [contributor, funder]