Base information for sequence characters of nucleotide strings containing modifications

The alphabet(), shortName() fullName() and nomenclature() functions return the letters, names and associated abbreviations for the type of ModString. alphabet() returns the normal letters and modification letters, whereas shortName(), fullName() and nomenclature() return results for modifications only.

shortName(x)

fullName(x)

nomenclature(x)

# S4 method for ModString
alphabet(x, baseOnly = FALSE)

# S4 method for ModStringSet
alphabet(x, baseOnly = FALSE)

# S4 method for ModString
shortName(x)

# S4 method for ModStringSet
shortName(x)

# S4 method for ModString
fullName(x)

# S4 method for ModStringSet
fullName(x)

# S4 method for ModString
nomenclature(x)

# S4 method for ModStringSet
nomenclature(x)

Arguments

x
a ModString or ModStringSet object

baseOnly
TRUE or FALSE (default): Should the result omit occurances of the letters N.-+?

Value

a character vector.

Examples

alphabet(ModDNAString())

#>  [1] "A"  "C"  "G"  "T"  "N"  "-"  "+"  "."  "p"  "δ"  "O"  "]"  "D"  "J"  "e" 
#> [16] "g"  "`"  "b"  "U"  "∝"  "π"  "I"  "7"  "6"  "3"  "2"  "1"  "8"  "∉"  "⊆" 
#> [31] "⊇"  "R"  "α"  "m"  "h"  "×"  "f"  "4"  "ν"  "X"  "'"  "κ"  "o"  "("  ")" 
#> [46] "η"  "a"  "⇓"  "⇑"  "\"" "√"  "/"  "≡"  "ζ"  "~" 

shortName(ModDNAString())

#>  [1] "5pT"    "3mT"    "O4meT"  "1mT"    "dhT"    "baseJ"  "5fU"    "5hmU"  
#>  [9] "dhpU"   "5caU"   "dU"     "DHdU"   "5nmU"   "dI"     "7mG"    "6mG"   
#> [17] "3mG"    "2mG"    "1mG"    "8oxoG"  "7a7dG"  "12eG"   "23eG"   "2,2mG" 
#> [25] "2ceG"   "5mC"    "5hmC"   "5gmC"   "5fC"    "5caC"   "4mC"    "34eC"  
#> [33] "3mC"    "3eC"    "2mC"    "1mC"    "9mA"    "7mA"    "6mA"    "4mA"   
#> [41] "3mA"    "1mA"    "6haA"   "2mA"    "2a6haA" "6,6mA"  "6ncmA" 

nomenclature(ModDNAString())

#>  [1] "5pT"    "3mT"    "O4meT"  "1mT"    "dhT"    "baseJ"  "5fU"    "5hmU"  
#>  [9] "dhpU"   "5caU"   "dU"     "DHdU"   "5nmU"   "dI"     "7mG"    "6mG"   
#> [17] "3mG"    "2mG"    "1mG"    "8oxoG"  "7a7dG"  "12eG"   "23eG"   "2,2mG" 
#> [25] "2ceG"   "5mC"    "5hmC"   "5gmC"   "5fC"    "5caC"   "4mC"    "34eC"  
#> [33] "3mC"    "3eC"    "2mC"    "1mC"    "9mA"    "7mA"    "6mA"    "4mA"   
#> [41] "3mA"    "1mA"    "6haA"   "2mA"    "2a6haA" "6,6mA"  "6ncmA"